refactor!: bundle microservices and dockerize
