+++ backend/user-service/src/model/user.types.ts
